Log all errors in `lib/error_handler`
Previously, the error handler would not log the original error message for the
case when something fails a Mongoose validation, since we replace those errors
with a custom message to avoid leaking private info to the client.

The simplest solution is to just log that original error before overwriting it.

Close #21.